+++ addons/ui/XEH_preStart.sqf
@@ -1,3 +1,3 @@
 #include "script_component.hpp"
 
-PREP(configureDisplayLoad);
+PREP(initDisplayInterrupt);
