+++ addons/ui/fnc_initDisplayInterrupt.sqf
